Fix compile errors and warnings on MSVC

+++ src/window/empire.c
@@ -376,11 +376,11 @@ void draw_trade_dots(const empire_object *trade_route, int x_draw_offset, int y_
     double x_factor = x_diff / dist;
     double y_factor = y_diff / dist;
     static const int dot_spacing = 15;
-    int num_dots = ceil(dist / dot_spacing);
+    int num_dots = (int) ceil(dist / dot_spacing);
     int image_id = trade_route->type == EMPIRE_OBJECT_LAND_TRADE_ROUTE ? assets_get_image_id("UI", "LandRouteDot") : assets_get_image_id("UI", "SeaRouteDot");
     for (int j = 0; j < num_dots; j++) {
-        int x = x_factor * j * 15 + start_x;
-        int y = y_factor * j * 15 + start_y;
+        int x = (int) (x_factor * j * 15 + start_x);
+        int y = (int) (y_factor * j * 15 + start_y);
         image_draw(image_id, x_draw_offset + x, y_draw_offset + y, COLOR_MASK_NONE, SCALE_NONE);
     }
 }
